requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-crispy-forms==1.8.0
django-multiselectfield==0.1.10
docopt==0.4.0
gunicorn==20.0.4
idna==2.8
mailchimp==2.0.9
numpy==1.17.3
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
